create update listing api route

+++ api/controllers/listing.controller.js
@@ -1,5 +1,6 @@
 import Listing from "../models/listing.model.js";
 import { errorHandler } from "../utils/error.js";
+import mongoose from "mongoose";
 
 export const createListing = async (req,res,next) => {
     try{
@@ -11,12 +12,17 @@ export const createListing = async (req,res,next) => {
 }
 
 export const deleteListing = async (req,res,next) => {
+   
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return next(errorHandler(400, 'Invalid Listing ID!'));
+    }
+
     const listing = await Listing.findById(req.params.id);
     if(!listing){
         return next(errorHandler(404 , 'Listing not found!'));
     }
     if(req.user.id !== listing.useRef.toString()){
-        return next(errorHandler(401 , 'You can onlt delete your own listings!'));
+        return next(errorHandler(401 , 'You can only delete your own listings!'));
     }
     try{
         await Listing.findByIdAndDelete(req.params.id);
@@ -25,3 +31,28 @@ export const deleteListing = async (req,res,next) => {
         next(error);
     }
 }
+
+export const updateListing = async (req, res, next) => {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return next(errorHandler(400, 'Invalid Listing ID!'));
+    }
+
+    const listing = await Listing.findById(req.params.id);
+    if (!listing) {
+      return next(errorHandler(404, 'Listing not found!'));
+    }
+    if (req.user.id !== listing.useRef.toString()) {
+      return next(errorHandler(401, 'You can only update your own listings!'));
+    }
+  
+    try {
+      const updatedListing = await Listing.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        { new: true }
+      );
+      res.status(200).json(updatedListing);
+    } catch (error) {
+      next(error);
+    }
+  };
